move: move SubPageHeader and test to ui dir
